pyproject.toml
[project]
name = "db-monitor"
version = "0.1.0"
description = "A comprehensive Flask application for monitoring and managing databases with SQL editor and real-time connection status"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-socketio>=5.3.0",
    "sqlalchemy>=2.0.0",
    "python-socketio>=5.9.0",
    "pymysql>=1.1.0",
    "psycopg2-binary>=2.9.0",
    "python-engineio>=4.7.0",
    "cryptography>=46.0.4",
    "gunicorn>=25.1.0",
    "eventlet>=0.40.4",
    "python-json-logger>=4.0.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-instrumentation-flask>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "opentelemetry-exporter-otlp>=1.39.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]

[tool.ruff]
line-length = 112
indent-width = 4

[tool.ruff.format]
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[project.optional-dependencies]
ldap = ["ldap3>=2.9.0"]

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
pythonpath = [
  "."
]
